backend/requirements.txt
fastapi
uvicorn[standard]
aiokafka
cryptography
aiofiles
databases[postgresql]
sqlalchemy
asyncpg
python-multipart
